backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "esgard",email = "pesgard@gmail.com"}
]
license = {text = "GPL"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyswip (>=0.3.3,<0.4.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-cloud-dialogflow (>=2.43.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
